refactor: enhance compiler analysis and parsing for linker remainder handling
